Final clean-up of ListOffsetArray.
NumpyArray is working (with __repr__ and such).

Introduce AtType.

Split up the initializer list that breaks Windows 32 to see which argument it is.

Try setting integer types explicitly for Windows 32 and 64 bit.

Start to write ListOffsetArray; we'll need slicing for that.

Also drop that Raw.cpp file.

First test of ListOffsetArray seems to be working well.