Start using RegularArray everywhere it needs to be used. (#24)
This check-in includes a commented-out attempt at `lower_getitem_nothing` in `awkward1._numba.array.numpyarray`. In the next PR, it will be gone. I concluded that there is no non-dependent type inference (where array dimensionality is part of the type but the length is not) for `__getitem__` of arrays for `NumpyArray` because an empty array would reduce the dimensionality by one yet return nothing, but reducing the dimensionality of a one-dimensional `NumpyArray` would return a scalar, which is not nothing. Whereas the C++ implementation could be fully transferred from a `ListArray` wrapper to a `RegularArray` wrapper (because it is dynamically typed, ironically), the Numba implementation could not (because it is not dynamically typed). The Numba implementation only wraps with `RegularArray` if the types are sound.

* Start using RegularArray everywhere it needs to be used.

* [skip ci] ListArray64 wrapper -> RegularArray wrapper partly works.

* [skip ci] ListArray64 wrapper -> RegularArray wrapper still only works partially.

* ListArray64 wrapper -> RegularArray wrapper now works

* RegularArray::tojson_part and fix EmptyArray::tojson_part.

* Centralize lower_getitem_tuple and lower_getitem_other.

* Avoid selecting types based on a value (length of array) in Content::getitem(Slice).

* getitem_tuple for Numba mirrors C++ as well as possible: if the 'outtpe' is 'NumpyArrayType', then you're going to have to 'lower_getitem_int'; 'lower_getitem_nothing' is impossible.

* ListArray and ListOffsetArray::getitem_next(array, not advanced) returns a RegularArray now (C++ and Numba).

* ListArray and ListOffsetArray::getitem_next(array) returns the same shape of RegularArrays as the index array.

* Now newaxis uses RegularArray, too.

* Ambiguous identities in ListArray are now refused with an error.

* [skip ci] writing awkward_identity*_from_listoffsetarray*.

* ListOffsetArray now uses awkward_identity*_from_listoffsetarray*.

* Addressed the FIXMEs in getitem.cpp; ready to merge PR.