Issue #50: IndexedArray::flatten for axis=0 (#81)
Added `IndexedArray::flatten` for `axis=0` (other cases not covered, just as they're not covered for the other arrays, either).

Also added `*::count` (a special-case reducer) for `axis=0` because I thought I needed it. In fact, I also added a version that doesn't depend on `axis` and returns an `Index64` because I thought I need that, too. As it turns out, I was misinterpreting the intended output, and I didn't actually need either `count` method. However, it's probably going to be needed to implement `getitem` slicing for jagged arrays, so I'm leaving it in.

The `*::count` reducer should be extended to handle any `axis` value as part of issue #51. The handling is the same as it would be for `*::flatten`, so it should be abstracted out into a helper function.

* [WIP] Issue #50: IndexedArray::flatten for axis=0

* Set up arrays for testing IndexedArray::flatten.

* Implemented most of IndexedArray::flatten, but IndexedOptionArray case is wrong.

* Implemented *::count so that it will be possible to implement IndexedOptionArray, but the latter isn't done yet.

* Tests for *::count.

* Fix 32-bit.

* Need to implement toindex64 to implement IndexedOptionArray::flatten.

* [skip ci] toindex64 isn't going to work.

* Replace toindex64 with count64 (specifically what we need right now).

* Finished IndexedArray::count64.

* [skip ci] Working on IndexedOptionArray::flatten (finally).

* IndexedOptionArray::flatten is done (and it didn't need counts, after all).

* Update signatures.

* It was missing some index types.