Make ak.Array a Pandas DType extension. (#92)
Awkward Arrays can be used as Pandas columns now. The tests now require `pandas>=0.24` and have probably been tested with version 0.25. As of this writing, Pandas 1.0 has not been released (except as a candidate), and Pandas 1.0 has not been tested.

Iteration over data in Matplotlib has been inhibited by the fact that Matplotlib calls `np.size` and `np.atleast_1d` on the arrays it's given, so we implemented `ak.size` and `ak.atleast_1d`, which raise `ValueErrors` if the arrays can't be cast as NumPy. At least it's better than grinding a computer to a halt as it tries to iterate over all the irregular items.

* [WIP] Make ak.Array a Pandas DType extension.

* The number is 92.

* ak.Array is a Pandas extension and DOES NOT 'import pandas' until 'dtype' is accessed.

* Add Pandas as a test-requirement. Let's see which Azure nodes fail.

* Skip py27 case and test_interesting. Added called_by_module(...) so that __array__ behavior can be more liberal when called by Pandas.

* Implemented ak.Array.nbytes.

* Stubs for all other Pandas methods.

* Implemented Content::deep_copy and connected it to ak.Array.copy.

* Do 'take' next.

* Organizing and generalizing the Python side.

* Implemented 'take' for Pandas.

* All Pandas tests pass.

* Matplotlib has been inhibited without a specialized check (just appropriate implementations of ak.size and ak.atleast_1d).