libawkward: add destructor definitions for some polymorphic classes
This is done to avoid emitting vtables in dependent libraries. The
weak-vtable warning will serve as a test for the future. It may not be
able to see some of the header-only classes. This commit covers
classes that need to be fixed in order for the build to succeed on
clang 7.1.0.

http://llvm.org/docs/CodingStandards.html#provide-a-virtual-method-anchor-for-classes-in-headers
https://gcc.gnu.org/onlinedocs/gcc-9.2.0/gcc/Vague-Linkage.html