IndexedArray.cpp: reorder template instantiations to fix dylib
There is an issue with instantiation of `IndexedArrayOf<int64_t, true>`:
```
0000000000055ec0 S typeinfo for awkward::IndexedArrayOf<int, false>
0000000000055f20 S typeinfo for awkward::IndexedArrayOf<int, true>
0000000000055ee0 S typeinfo for awkward::IndexedArrayOf<unsigned int, false>
0000000000055f00 S typeinfo for awkward::IndexedArrayOf<long long, false>
0000000000055f40 S typeinfo for awkward::IndexedArrayOf<long long, true>
```
but it gets linked to a private symbol (note lowercase "s")
```
0000000000266c20 S typeinfo for awkward::IndexedArrayOf<int, true>
0000000000266be0 S typeinfo for awkward::IndexedArrayOf<unsigned int, false>
0000000000266c00 S typeinfo for awkward::IndexedArrayOf<long long, false>
0000000000266c40 s typeinfo for awkward::IndexedArrayOf<long long, true>
```
That renders libawkward.dylib unusable.

Unfortunately, I am not able to craft a minimal reproducing example
for this, but it reliably fails on darwin.

This may be a bug in linker. This reproduces with:
```
@(#)PROGRAM:ld  PROJECT:ld64-450.3
```