Feature: add layout transformers
The existing recursively_apply implementation is designed for one-off
layout modifications. There are a number of places where we might want
to visit multiple contents in a layout, and transform more than one of
them. This commit contains a new `LayoutTransformer` (and associated
`FunctionalTransformer` which implements a getfunction-like API) to
support this use case.

Classes were used rather than closures in order to cleanly handle the
mutual dependency of the transformer and the generic transformer routines
upon one another.