Documentation: WIP on `ak.ravel` vs `ak.flatten(axis="records")`
This might be a bit overzealous on using `ak.ravel` vs `ak.flatten(..., axis="records")`. Long term, we want `ak.flatten` to have `axis=None` behave as `axis="records"`, which will be a more ergonomic API. It doesn't seem obvious why `axis="records"` flattens entirely.
This is part of a bigger discussion as to whether `ak.ravel` should be promoted as a function. It is likely that `ak.ravel` should be "discouraged" in favour of explicit flattening, in order to avoid losing all the structure.

For now, though, `axis="records"` is explicit, and this commit uses it in favor of `ravel`