fix: ensure that `__copy__` and `__deepcopy__` are enabled. (#1695)
* fix: ensure that  and  are enabled.

* Implemented all the (missing) copy/deepcopy methods.

* refactor: use private parameters where possible

* fix: use `_length` not `_zeros_length` in `RegularArray.copy()``

NB. I did think about introducing a `zeros_length` attribute, but I am concerned that we would need to take a great deal of care in handling this parameter for no good cause. Instead, we should make it such that the `length` parameter is the only externally visible one.

* refactor: replace `Content.__deepcopy__` with abstract method

This will ensure we *have* to implement this routine. Although we could use a dynamic deepcopy, it opens us up to more bugs (explicit vs implicit) and hides some of the copying logic.

* refactor: remove unneeded optional in `__deepcopy__` signature.

* fix: use `copy.deepcopy` to dispatch copying

This avoids handling deep copying in two places

* fix: `deepcopy(behavior)` in `Array/Record.__deepcopy__`

* fix: use `copy.deepcopy` in tests

Co-authored-by: Angus Hollands <goosey15@gmail.com>