fix: ak.from_numpy should fail on zero-dimensional arrays. (#3161)
* Fixing policy issue 1057

At this point it's a matter of consistency. Now *both* `ak.Array()` and `ak.from_numpy()` will throw a TypeError when passed either a python scalar, a numpy numeric value like `np.float64(5.2)`, **or** a numpy zero-dimensional-array like `np.array(5.2)`. But in the latter case there is an option to allow this, which is necessary for lots of internal functions, by providing a value for `primitive_policy` other than "error."

Prior to this patch the first test, ak.Array(), was already passing without modifications.
But the second test was not -- ak.from_numpy().
Fix is in _layout.py. No policy options are passed to _layout.from_arraylib().

* Adding/passing a primitive_policy kwarg

From from_numpy, from_cupy, from_jax, and from_dlpack
To from_arraylib

---------

Co-authored-by: Ianna Osborne <ianna.osborne@cern.ch>